[TNT-3920] use the source branch values
For pull request events, by default `GITHUB_REF` and `GITHUB_SHA` point to the merge branch and merge commit SHA respectively,
not the PR's source branch and head commit SHA.
This can lead to confusion and incorrect behavior in workflows that rely on these values (for instance, in the Deploy app UI).

This change modifies the action to automatically use:
- The PR's head commit SHA instead of the merge commit SHA.
- The PR's source branch name instead of the merge branch.

A #major version has been added in the CHANGELOG, as it alters the default behavior of the action.